refactor(dsh-mcp-manager): 阶段6批次4——api/ + bootstrap/ 域落位（routes* + apply* + interface.ts，Part of #664）
git mv 至目标目录（v3 §二）：
- api/：routes/routes-controllers/routes-helpers
- bootstrap/：apply/apply-config/apply-runtime/apply-services/apply-guidance
每目录新增 interface.ts 门面（D10）；index.ts 汇聚 re-export 切
bootstrap/api 门面；service-contract 静态扫描路径同步
src/bootstrap/apply-services.ts；shared 相对路径按新深度修正。
纯搬移，零行为变更（lib 导出面不变）。

Related issue #664
