+++ packages/dsh-mcp-manager/src/api/routes-helpers.ts
