+++ packages/dsh-mcp-manager/src/bootstrap/apply-guidance.ts
